fix tests: isolate gpg and git more from the base system
The test suite provides some git configuration, additional
custom git configuration should not cause test cases to fail.

GPG has a maximum length for the path to the GPG agent socket,
Copying the entire gpg directory to a temporary directory
ensures that the path is not too long when the repo is cloned into a somewhat
nested directory structure.